backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
anthropic==0.25.0
python-docx==0.8.11
PyPDF2==3.0.1
docx2txt==0.8
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
